sdk/python: move services into services package
